+++ rust/operator-binary/src/airflow_controller.rs
@@ -67,18 +67,17 @@ use stackable_operator::{
 use strum::{EnumDiscriminants, IntoStaticStr};
 
 use crate::{
-    controller::build::config_map,
+    controller::{AirflowRoleGroupConfig, ValidatedCluster, build::config_map},
     controller_commons::{self, CONFIG_VOLUME_NAME, LOG_CONFIG_VOLUME_NAME, LOG_VOLUME_NAME},
     crd::{
-        self, APP_NAME, AirflowClusterStatus, AirflowConfig, AirflowConfigOverrides,
-        AirflowExecutor, AirflowExecutorCommonConfiguration, AirflowRole, CONFIG_PATH, Container,
-        ExecutorConfig, HTTP_PORT, HTTP_PORT_NAME, LISTENER_VOLUME_DIR, LISTENER_VOLUME_NAME,
-        LOG_CONFIG_DIR, METRICS_PORT, METRICS_PORT_NAME, OPERATOR_NAME, STACKABLE_LOG_DIR,
-        TEMPLATE_LOCATION, TEMPLATE_NAME, TEMPLATE_VOLUME_NAME,
+        self, APP_NAME, AirflowClusterStatus, AirflowConfigOverrides, AirflowExecutor,
+        AirflowExecutorCommonConfiguration, AirflowRole, CONFIG_PATH, Container, ExecutorConfig,
+        HTTP_PORT, HTTP_PORT_NAME, LISTENER_VOLUME_DIR, LISTENER_VOLUME_NAME, LOG_CONFIG_DIR,
+        METRICS_PORT, METRICS_PORT_NAME, OPERATOR_NAME, STACKABLE_LOG_DIR, TEMPLATE_LOCATION,
+        TEMPLATE_NAME, TEMPLATE_VOLUME_NAME,
         authentication::{
             AirflowAuthenticationClassResolved, AirflowClientAuthenticationDetailsResolved,
         },
-        authorization::AirflowAuthorizationResolved,
         build_recommended_labels,
         internal_secret::{
             FERNET_KEY_SECRET_KEY, INTERNAL_SECRET_SECRET_KEY, JWT_SECRET_SECRET_KEY,
@@ -106,106 +105,6 @@ pub const AIRFLOW_FULL_CONTROLLER_NAME: &str =
 pub struct Ctx {
     pub client: stackable_operator::client::Client,
     pub operator_environment: OperatorEnvironmentOptions,
-}
-
-/// Per-role configuration extracted during validation.
-#[derive(Clone, Debug)]
-pub struct ValidatedRoleConfig {
-    pub pdb: Option<stackable_operator::commons::pdb::PdbConfig>,
-    pub listener_class: Option<String>,
-    pub group_listener_name: Option<String>,
-}
-
-/// Per-rolegroup configuration: the merged CRD config plus overrides.
-///
-/// This is the generic [`stackable_operator::v2::role_utils::RoleGroupConfig`]: the merged config
-/// fragment in `config`, the typed `config_overrides` (role-group merged over role) and the merged
-/// `env_overrides`/`cli_overrides`/`pod_overrides`. The config overrides are kept typed
-/// ([`AirflowConfigOverrides`]) and flattened into the rendered config file later, in the build step.
-pub type AirflowRoleGroupConfig = stackable_operator::v2::role_utils::RoleGroupConfig<
-    AirflowConfig,
-    stackable_operator::v2::role_utils::GenericCommonConfig,
-    AirflowConfigOverrides,
->;
-
-/// Cluster-wide configuration that applies to every role and role group.
-///
-/// Carries the dereferenced external references, so every downstream build step reads them from
-/// here rather than from the raw cluster object.
-#[derive(Clone, Debug)]
-pub struct ValidatedClusterConfig {
-    pub executor: AirflowExecutor,
-    pub authentication_config: AirflowClientAuthenticationDetailsResolved,
-    pub authorization_config: AirflowAuthorizationResolved,
-}
-
-/// The validated cluster: proves that config merging succeeded for every role and
-/// role group before any resources are created.
-#[derive(Clone, Debug)]
-pub struct ValidatedCluster {
-    /// `ObjectMeta` carrying `name`, `namespace` and `uid`, captured during validation, so this
-    /// struct can stand in as the owner [`Resource`] for child objects.
-    metadata: ObjectMeta,
-    pub image: ResolvedProductImage,
-    pub cluster_config: ValidatedClusterConfig,
-    pub role_groups: BTreeMap<AirflowRole, BTreeMap<String, AirflowRoleGroupConfig>>,
-    pub role_configs: BTreeMap<AirflowRole, ValidatedRoleConfig>,
-}
-
-impl ValidatedCluster {
-    pub fn new(
-        airflow: &v1alpha2::AirflowCluster,
-        image: ResolvedProductImage,
-        cluster_config: ValidatedClusterConfig,
-        role_groups: BTreeMap<AirflowRole, BTreeMap<String, AirflowRoleGroupConfig>>,
-        role_configs: BTreeMap<AirflowRole, ValidatedRoleConfig>,
-    ) -> Self {
-        Self {
-            // Capture only the identity fields needed to own child objects.
-            metadata: ObjectMeta {
-                name: Some(airflow.name_any()),
-                namespace: airflow.namespace(),
-                uid: airflow.uid(),
-                ..ObjectMeta::default()
-            },
-            image,
-            cluster_config,
-            role_groups,
-            role_configs,
-        }
-    }
-}
-
-/// Lets [`ValidatedCluster`] stand in for the raw [`v1alpha2::AirflowCluster`] when building owner
-/// references and metadata for child objects. Kind/group/version are delegated to the CRD; the
-/// `metadata` (name, namespace, uid) is captured during validation.
-impl Resource for ValidatedCluster {
-    type DynamicType = <v1alpha2::AirflowCluster as Resource>::DynamicType;
-    type Scope = <v1alpha2::AirflowCluster as Resource>::Scope;
-
-    fn kind(dt: &Self::DynamicType) -> std::borrow::Cow<'_, str> {
-        v1alpha2::AirflowCluster::kind(dt)
-    }
-
-    fn group(dt: &Self::DynamicType) -> std::borrow::Cow<'_, str> {
-        v1alpha2::AirflowCluster::group(dt)
-    }
-
-    fn version(dt: &Self::DynamicType) -> std::borrow::Cow<'_, str> {
-        v1alpha2::AirflowCluster::version(dt)
-    }
-
-    fn plural(dt: &Self::DynamicType) -> std::borrow::Cow<'_, str> {
-        v1alpha2::AirflowCluster::plural(dt)
-    }
-
-    fn meta(&self) -> &ObjectMeta {
-        &self.metadata
-    }
-
-    fn meta_mut(&mut self) -> &mut ObjectMeta {
-        &mut self.metadata
-    }
 }
 
 #[derive(Snafu, Debug, EnumDiscriminants)]
